Push from fossil repo [a2ba515899]
Added project lzjson to github repository. First commit is the base
code which compiles (at least under Windows) with a test.c and test.txt
to test against. The project is incomplete as testing of continuances
and other features is not done yet.

Nor is any documentation. Soon I suppose.